+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 * Introduce `consLensBE32` for efficient serialization of wire protocols
   that require prefixing a payload with its length.
-* Add `int64BE` as a convenience.
+* Add `int{16,32,64}BE` and `int{16,32,64LE}` as conveniences.
 * Add little-endian encoding functions for `Word16`, `Word32`, and `Word64`.
 * Add big-endian and little-endian functions for copying a
-  `PrimArray` of numbers into a builder.
+  `PrimArray` of numbers (both signed and unsigned) into a builder.
 * Add `flush`, `copy`, and `insert` for better control when
   converting byte sequences to builders.